+++ frontend/src/app/core/connection-lost-modal/open-connection-lost-modal.ts
@@ -1,0 +1,10 @@
+import type { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { ConnectionLostModalComponent } from './connection-lost-modal.component';
+
+export function openConnectionLostModal(ngbModalService: NgbModal) {
+    ngbModalService.open(ConnectionLostModalComponent, {
+        size: 'xl',
+        beforeDismiss: () => false,
+        centered: true,
+    });
+}